-- fixed various bugs related to the new way in which the new xcode compiler is reporting errors
git-svn-id: https://svn.rtmpd.com/crtmpserver/trunk@759 01689507-d990-df11-9d7d-0017a411d416

+++ sources/thelib/include/protocols/rtmp/amftypes.h
@@ -1,18 +1,18 @@
-/* 
+/*
  *  Copyright (c) 2010,
  *  Gavriloaie Eugen-Andrei (shiretu@gmail.com)
- *  
+ *
  *  This file is part of crtmpserver.
  *  crtmpserver is free software: you can redistribute it and/or modify
  *  it under the terms of the GNU General Public License as published by
  *  the Free Software Foundation, either version 3 of the License, or
  *  (at your option) any later version.
- *  
+ *
  *  crtmpserver is distributed in the hope that it will be useful,
  *  but WITHOUT ANY WARRANTY; without even the implied warranty of
  *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  *  GNU General Public License for more details.
- *  
+ *
  *  You should have received a copy of the GNU General Public License
  *  along with crtmpserver.  If not, see <http://www.gnu.org/licenses/>.
  */
@@ -44,32 +44,32 @@
 	AMF0_P_NS_Int32 = 0xf1,
 	AMF0_P_uninitialized = 0xff*/
 
-#define AMF0_NUMBER         0x00
-#define AMF0_BOOLEAN        0x01
-#define AMF0_SHORT_STRING   0x02
-#define AMF0_OBJECT         0x03
-#define AMF0_NULL           0x05
-#define AMF0_UNDEFINED      0x06
-#define AMF0_MIXED_ARRAY    0x08
-#define AMF0_ARRAY          0x0a
-#define AMF0_TIMESTAMP      0x0b
-#define AMF0_LONG_STRING    0x0c
-#define AMF0_TYPED_OBJECT	0x10
-#define AMF0_AMF3_OBJECT    0x11
+#define AMF0_NUMBER         ((uint8_t)0x00)
+#define AMF0_BOOLEAN        ((uint8_t)0x01)
+#define AMF0_SHORT_STRING   ((uint8_t)0x02)
+#define AMF0_OBJECT         ((uint8_t)0x03)
+#define AMF0_NULL           ((uint8_t)0x05)
+#define AMF0_UNDEFINED      ((uint8_t)0x06)
+#define AMF0_MIXED_ARRAY    ((uint8_t)0x08)
+#define AMF0_ARRAY          ((uint8_t)0x0a)
+#define AMF0_TIMESTAMP      ((uint8_t)0x0b)
+#define AMF0_LONG_STRING    ((uint8_t)0x0c)
+#define AMF0_TYPED_OBJECT	((uint8_t)0x10)
+#define AMF0_AMF3_OBJECT    ((uint8_t)0x11)
 
-#define AMF3_UNDEFINED      0x00
-#define AMF3_NULL           0x01
-#define AMF3_FALSE          0x02
-#define AMF3_TRUE           0x03
-#define AMF3_INTEGER        0x04
-#define AMF3_DOUBLE         0x05
-#define AMF3_STRING         0x06
-#define AMF3_XMLDOC         0x07
-#define AMF3_DATE           0x08
-#define AMF3_ARRAY          0x09
-#define AMF3_OBJECT         0x0a
-#define AMF3_XML            0x0b
-#define AMF3_BYTEARRAY      0x0c
+#define AMF3_UNDEFINED      ((uint8_t)0x00)
+#define AMF3_NULL           ((uint8_t)0x01)
+#define AMF3_FALSE          ((uint8_t)0x02)
+#define AMF3_TRUE           ((uint8_t)0x03)
+#define AMF3_INTEGER        ((uint8_t)0x04)
+#define AMF3_DOUBLE         ((uint8_t)0x05)
+#define AMF3_STRING         ((uint8_t)0x06)
+#define AMF3_XMLDOC         ((uint8_t)0x07)
+#define AMF3_DATE           ((uint8_t)0x08)
+#define AMF3_ARRAY          ((uint8_t)0x09)
+#define AMF3_OBJECT         ((uint8_t)0x0a)
+#define AMF3_XML            ((uint8_t)0x0b)
+#define AMF3_BYTEARRAY      ((uint8_t)0x0c)
 
 #define AMF_CHECK_BOUNDARIES(x,y) \
 do { \
